// plugins/docusaurus-plugin-typedoc-api/src/types.ts
export interface DocusaurusPluginTypeDocApiOptions
  extends Omit<VersionsOptions, "disableVersioning" | "includeCurrentVersion"> {
  includeApiDir?: boolean;
  banner?: string;
  breadcrumbs?: boolean;
  changelogName?: string;
  changelogs?: boolean;
  gitRefName?: string;
  debug?: boolean;
  exclude?: string[];
  id?: string;
  minimal?: boolean;
  packageJsonName?: string;
  packages: (PackageConfig | string)[];
  projectRoot: string;
  readmeName?: string;
  readmes?: boolean;
  removeScopes?: string[];
  sortPackages?: (
    a: PackageReflectionGroup,
    d: PackageReflectionGroup
  ) => number;
  sortSidebar?: (a: string, d: string) => number;
  tsconfigName?: string;
  typedocOptions?: Partial<
    Pick<
      TypeDocOptions,
      | "commentStyle"
      | "disableSources"
      | "emit"
      | "excludeExternals"
      | "excludeInternal"
      | "excludeNotDocumented"
      | "excludePrivate"
      | "excludeProtected"
      | "excludeTags"
      | "externalPattern"
      | "logger"
      | "logLevel"
      | "sort"
      | "treatWarningsAsErrors"
      | "validation"
    >
  >;

  // Versioning, based on Docusaurus
  disableVersioning?: boolean;
  includeCurrentVersion?: boolean;
  routeBasePath?: string;

  apiDocs: PluginOptions;
}

// CONFIG

export interface PackageEntryConfig {
  label: string;
  path: string;
}

export interface PackageConfig {
  path: string; // Folder relative to project root
  entry?: Record<string, PackageEntryConfig | string> | string;
  slug?: string;
}

export interface ResolvedPackageConfig {
  entryPoints: Record<string, PackageEntryConfig>;
  packagePath: string;
  packageSlug: string;
  packageName: string;
  packageVersion: string;
}
// ...
export type SidebarItem = PropSidebarItem;

export interface ApiOptions {
  banner: string;
  breadcrumbs: boolean;
  gitRefName: string;
  minimal: boolean;
  pluginId: string;
  scopes: string[];
}

export interface TOCItem {
  readonly value: string;
  readonly id: string;
  readonly level: number;
}

// REFLECTIONS

export interface PackageReflectionGroupEntry {
  index: boolean;
  label: string;
  reflection: JSONOutput.ProjectReflection;
  urlSlug: string;
}

export interface PackageReflectionGroup {
  entryPoints: PackageReflectionGroupEntry[];
  packageName: string;
  packageVersion: string;
  changelogPath?: string;
  readmePath?: string;
}